对算法加入基本的异常输入处理功能 根据输入划分原则丰富了测试用例
